product/requirements.txt
Flask==2.0.3
Werkzeug==2.0.3
flask-pymongo==2.3.0
flask-jwt-extended==4.3.1
python-dotenv==0.19.2
requests==2.26.0
pymongo==3.12.1
pymongo[srv]